:zap: CalculatorToVoteCountPercentage 함수 Util 이동
